+++ src/Lykke.Job.TransactionHandler.Core/Services/Messages/Email/ContentGenerator/MessagesData/TransferCompletedData.cs
@@ -3,9 +3,9 @@
     public class TransferCompletedData : IEmailMessageData
     {
         public string ClientName { get; set; }
-        public double AmountFiat { get; set; }
-        public double AmountLkk { get; set; }
-        public double Price { get; set; }
+        public decimal AmountFiat { get; set; }
+        public decimal AmountLkk { get; set; }
+        public decimal Price { get; set; }
         public string AssetId { get; set; }
         public string SrcBlockchainHash { get; set; }
         public string MessageId()
