+++ src/Lykke.Job.TransactionHandler.Services/Fee/FeeCalculationService.cs
@@ -17,13 +17,13 @@ namespace Lykke.Job.TransactionHandler.Services.Fee
             _assetsServiceWithCache = assetsServiceWithCache ?? throw new ArgumentNullException(nameof(assetsServiceWithCache));
         }
 
-        public async Task<double> GetAmountNoFeeAsync(double initialAmount, string assetId, List<Core.Contracts.Fee> fees)
+        public async Task<decimal> GetAmountNoFeeAsync(decimal initialAmount, string assetId, List<Core.Contracts.Fee> fees)
         {
             var asset = await _assetsServiceWithCache.TryGetAssetAsync(assetId);
 
             var feeTransfer = fees?.FirstOrDefault()?.Transfer;
 
-            var feeAmount = (feeTransfer?.Volume ?? 0d).TruncateDecimalPlaces(asset.Accuracy, true);
+            var feeAmount = (feeTransfer?.Volume ?? 0).TruncateDecimalPlaces(asset.Accuracy, true);
 
             return  initialAmount > 0 ? initialAmount - feeAmount : initialAmount + feeAmount;
         }
